refactor: auth 패키지의 Bean 등록을 Component Scan에서 Java Configuration로 변경
