+++ src/main/java/app/auth/LoginService.java
@@ -2,7 +2,6 @@ package app.auth;
 
 import org.springframework.stereotype.Service;
 
-
 @Service
 public class LoginService {
     private LoginDao loginDao;
