+++ src/main/java/app/nextstep/RoomEscapeApplication.java
@@ -1,4 +1,4 @@
-package app;
+package app.nextstep;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
