+++ src/test/java/app/auth/JwtTokenProviderTest.java
@@ -1,6 +1,6 @@
 package app.auth;
 
-import app.auth.JwtTokenProvider;
+import app.nextstep.RoomEscapeApplication;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("JwtTokenProvider 학습 테스트")
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes = RoomEscapeApplication.class)
 class JwtTokenProviderTest {
     @Autowired
     private JwtTokenProvider jwtTokenProvider;
